requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.1
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
idna==3.10
infobip-api-python-sdk==5.0.1
multidict==6.1.0
pillow==11.0.0
propcache==0.2.0
pydantic==1.10.19
PyJWT==2.9.0
regex==2022.3.15
requests==2.31.0
sqlparse==0.5.1
twilio==9.3.6
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
